requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
numpy
pandas
scikit-learn
joblib
matplotlib
seaborn
mysql-connector-python
